Server/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
openai==0.27.8
tiktoken==0.4.0
Flask-Migrate==4.0.4
Mako==1.2.4
alembic==1.12.0
psycopg2==2.9.7
redis==5.0.0
gunicorn==21.2.0
Flask-Limiter==3.5.0
amplitude-analytics==1.1.3